lib: sbi: Support Zcheripurecap ABI in system_opcode_insn
The `struct sbi_trap_regs` pointer is a CHERI capability pointer rather than a integer pointer
on Zcheripurecap ABI/ Change the pointer casting from `unsigned long` to `uintptr_t`

+++ include/sbi/riscv_encoding.h
@@ -997,7 +997,7 @@
 	(SHIFT_RIGHT((insn), (pos) - LOG_REGBYTES) & REG_MASK)
 
 #define REG_PTR(insn, pos, regs)	\
-	(ulong *)((ulong)(regs) + REG_OFFSET(insn, pos))
+	(uintptr_t *)((uintptr_t)(regs) + REG_OFFSET(insn, pos))
 
 #define GET_RM(insn)			(((insn) >> 12) & 7)
